requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
FLAML==2.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
numpy==1.26.4
ollama==0.1.6
openai==1.12.0
packaging==23.2
pyautogen==0.2.14
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
